+++ src/base/main/dabradio.h
@@ -354,7 +354,7 @@ public slots:
   void slot_show_frame_errors(i32 iFrameErrors);
   void slot_show_rs_errors(i32 iRsErrors);
   void slot_show_aac_errors(i32 iAacErrors);
-  void slot_show_fic_status(i32 iSuccessPercent, f32 iBER);
+  void slot_show_fic_status(f32 iSuccessPercent, f32 iBER);
   void slot_show_rs_corrections(i32 iC, i32 iEc);
 
   // Signal Quality and Measurements
